+++ src/program/reflection.rs
@@ -33,7 +33,7 @@ pub struct Uniform {
 /// Information about a uniform block (except its name).
 #[derive(Debug, Clone)]
 pub struct UniformBlock {
-    /// The binding point of the block.
+    /// Indentifier of the block.
     ///
     /// This is internal information, you probably don't need to use it.
     pub binding: i32,
@@ -372,9 +372,9 @@ pub unsafe fn reflect_uniform_blocks(ctxt: &mut CommandContext, program: Handle)
         };
 
         // binding point for this block
-        let mut binding: gl::types::GLint = mem::uninitialized();
+        /*let mut binding: gl::types::GLint = mem::uninitialized();
         ctxt.gl.GetActiveUniformBlockiv(program, block_id as gl::types::GLuint,
-                                        gl::UNIFORM_BLOCK_BINDING, &mut binding);
+                                        gl::UNIFORM_BLOCK_BINDING, &mut binding);*/
 
         // number of bytes
         let mut block_size: gl::types::GLint = mem::uninitialized();
@@ -439,7 +439,7 @@ pub unsafe fn reflect_uniform_blocks(ctxt: &mut CommandContext, program: Handle)
 
         // finally inserting into the blocks list
         blocks.insert(name, UniformBlock {
-            binding: binding as i32,
+            binding: block_id as i32,
             size: block_size as usize,
             layout: introspection_output_to_layout(members),
         });
@@ -669,13 +669,13 @@ pub unsafe fn reflect_shader_storage_blocks(ctxt: &mut CommandContext, program: 
 
     for block_id in (0 .. active_blocks) {
         // getting basic infos
-        let (name_len, num_variables, binding, total_size) = {
-            let mut output: [gl::types::GLint; 4] = mem::uninitialized();
-            ctxt.gl.GetProgramResourceiv(program, gl::SHADER_STORAGE_BLOCK, block_id, 4,
+        let (name_len, num_variables, total_size) = {
+            let mut output: [gl::types::GLint; 3] = mem::uninitialized();
+            ctxt.gl.GetProgramResourceiv(program, gl::SHADER_STORAGE_BLOCK, block_id, 3,
                                          [gl::NAME_LENGTH, gl::NUM_ACTIVE_VARIABLES,
-                                          gl::BUFFER_BINDING, gl::BUFFER_DATA_SIZE].as_ptr(), 4,
+                                          gl::BUFFER_DATA_SIZE].as_ptr(), 3,
                                          ptr::null_mut(), output.as_mut_ptr() as *mut _);
-            (output[0] as usize, output[1] as usize, output[2], output[3] as usize)
+            (output[0] as usize, output[1] as usize, output[2] as usize)
         };
 
         // getting the name of the block
@@ -732,7 +732,7 @@ pub unsafe fn reflect_shader_storage_blocks(ctxt: &mut CommandContext, program: 
 
         // finally inserting into the blocks list
         blocks.insert(name, UniformBlock {
-            binding: binding as i32,
+            binding: block_id as i32,
             size: total_size,
             layout: introspection_output_to_layout(members),
         });